+++ CS485_Bank_newb2710/01_Bank/CheckingAccount.h
@@ -12,7 +12,7 @@
 class CheckingAccount : public IAccount {
 	public:
 		CheckingAccount();
-		CheckingAccount(int acctNum, long long amt, float interestRate, IFee* pcTheFee);
+		CheckingAccount(int acctNum, long long amt, double interestRate, IFee* pcTheFee);
 		~CheckingAccount();
 	private:
